fix: rename json file to avoir conflict
